Update README.md
The computer’s main memory, also known as Random Access Memory (RAM), is an addressable sequence of n-bit registers, each designed to hold an n-bit value. In this project you will build a RAM module utilizing the skills and knowledge from the labs. This involves two main issues: (i) how to use gate logic to store bits persistently, over time, and (ii) how to use gate logic to locate the memory register (“using the address”) on which we wish to operate. In addition, you will build several functions that are constructed with combinational and sequential logic design elements.

Instructions

Download the collateral provided on Canvas and use the skeletal files provided. Do not change the name of the files.

Add your full name and UIN to the introductory comment present in each .hdl file.

Implement the .hdl for each chip.

Test the basic chips using the completed .tst and .cmp files provided.

Chips